+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ### Latest
--  Adding more general ratio_group argument to puma.Histogram [#67](https://github.com/umami-hep/puma/pull/67)
+- Change default *bb*-jets colour to dark red and vlines to black [#69](https://github.com/umami-hep/puma/pull/69)
+- Adding more general `ratio_group` argument to `puma.Histogram` [#67](https://github.com/umami-hep/puma/pull/67)
 - Adding `calc_separation()` to `puma.metrics`, which allows to calculate the separation between two distributions [#27](https://github.com/umami-hep/puma/pull/27)
 - Adding Zenodo link