+++ changelog.md
@@ -2,7 +2,7 @@
 
 ### Latest
 
-- Adding atlas_tag_outside [#71](https://github.com/umami-hep/puma/pull/71)
+- Adding atlas_tag_outside and change default for `atlas_second_tag` [#71](https://github.com/umami-hep/puma/pull/71)
 - Change default *bb*-jets colour to dark red and vlines to black [#69](https://github.com/umami-hep/puma/pull/69)
 - Adding more general `ratio_group` argument to `puma.Histogram` [#67](https://github.com/umami-hep/puma/pull/67)
 - Adding `calc_separation()` to `puma.metrics`, which allows to calculate the separation between two distributions [#27](https://github.com/umami-hep/puma/pull/27)