+++ changelog.md
@@ -2,7 +2,8 @@
 
 ### Latest
 
-- `ratio_group` can no longer be set via `flavour` argument [#74](https://github.com/umami-hep/puma/pull/74)
+- `ratio_group` in `puma.Histogram` objects can no longer be set via `flavour` argument [#74](https://github.com/umami-hep/puma/pull/74)
+
 - Adding example for `plt.show` replacement + adding theme switcher button to docs [#72](https://github.com/umami-hep/puma/pull/72)
 - Adding `atlas_tag_outside` and change default for `atlas_second_tag` [#71](https://github.com/umami-hep/puma/pull/71)
 - Change default *bb*-jets colour to dark red and vlines to black [#69](https://github.com/umami-hep/puma/pull/69)